Break out components, utils, and test-utils into separate directories + typechecking
